# nodejs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Szolgáltatás orientált programozás beadandó - Györkis Tamás",
    "description": "Az API-ban egyszerű, 10x10-es pixelart rajzok tárolhatók, kérdezhetők le, illetve frissíthetők, törölhetők. Emellett a gyakorlaton írt API-hoz is hozzáfér, le tudja kérdezni az ott tárolt egyes felhasználók jegyeit.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "tags": [
    {
      "name": "Rajzok",
      "description": "Rajzok lekérése, frissítése, törlése"
    },
    {
      "name": "PHP API",
      "description": "Hozzáférés a PHP API-hoz"
    },
    {
      "name": "Felhasználó",
      "description": "Regisztráció, bejelentkezés"
    }
  ],
  "paths": {
    "/drawings": {
      "get": {
        "tags": [
          "Rajzok"
        ],
        "summary": "Rajzok lekérdezése",
        "description": "Összes rajz adatainak lekérdezése",
        "operationId": "getAllDrawing",
        "responses": {
          "200": {
            "description": "Sikeres lekérdezés",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllDrawings"
                }
              }
            }
          },
          "400": {
            "description": "Nem adott meg azonosítót"
          },
          "404": {
            "description": "Rajz nem található"
          }
        }
      }
    },
    "/drawing/{id}": {
      "get": {
        "tags": [
          "Rajzok"
        ],
        "summary": "Egy rajz lekérdezése",
        "description": "Egy rajz adatainak lekérdezése",
        "operationId": "getDrawing",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "A keresett rajz azonosítója",
            "required": true,
            "explode": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sikeres lekérdezés",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SingleDrawing"
                }
              }
            }
          },
          "400": {
            "description": "Nem adott meg azonosítót"
          },
          "404": {
            "description": "Rajz nem található"
          }
        }
      },
      "put": {
        "tags": [
          "Rajzok"
        ],
        "summary": "Rajz frissítése",
        "description": "Meglévő rajz frissítése.",
        "operationId": "updateDrawing",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "A keresett rajz azonosítója",
            "required": true,
            "explode": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "description": "Rajz frissítése",
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDrawing"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Sikeres frissítés",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateDrawingResponse"
                }
              }
            }
          },
          "403": {
            "description": "Nincs hozzáférése"
          }
        }
      }
    },
    "/drawing": {
      "post": {
        "tags": [
          "Rajzok"
        ],
        "summary": "Új rajz",
        "description": "Új rajz felvitele a rendszerbe.",
        "operationId": "addDrawing",
        "requestBody": {
          "description": "Új rajz létrehozása",
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/NewDrawing"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Sikeres létrehozás",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewDrawingSuccess"
                }
              }
            }
          },
          "403": {
            "description": "Nincs hozzáférése"
          }
        }
      }
    },
    "/drawing/{token}/{id}": {
      "delete": {
        "tags": [
          "Rajzok"
        ],
        "summary": "Rajz törlése",
        "description": "Meglévő rajz törlése.",
        "operationId": "deleteDrawing",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "description": "A keresett rajz azonosítója",
            "required": true,
            "explode": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Token",
            "required": true,
            "explode": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sikeres törlés",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteDrawingResponse"
                }
              }
            }
          },
          "403": {
            "description": "Nincs hozzáférése"
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": [
          "Felhasználó"
        ],
        "summary": "Bejelentkezés",
        "description": "Bejelentkezés és token kérése",
        "operationId": "loginUser",
        "requestBody": {
          "description": "Token lekérése",
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        }
      }
    },
    "/register": {
      "post": {
        "tags": [
          "Felhasználó"
        ],
        "summary": "Regisztráció",
        "description": "Felhasználó regisztrálása",
        "operationId": "createUser",
        "requestBody": {
          "description": "Regisztráció",
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponse"
                }
              }
            }
          }
        }
      }
    },
    "/php": {
      "get": {
        "tags": [
          "PHP API"
        ],
        "summary": "PHP api elérése",
        "description": "PHP api elérése, jegyek lekérdezése",
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "description": "Felhasználónév a PHP API-ban",
            "required": true,
            "explode": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "password",
            "in": "query",
            "description": "Jelszó a PHP API-ban",
            "required": true,
            "explode": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sikeres lekérdezés",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PHPResponse"
                }
              }
            }
          },
          "400": {
            "description": "Nem adott meg azonosítót"
          },
          "404": {
            "description": "Felhasználó nem található"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "PHPResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "integer",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "Listing grades"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GradeItem"
            }
          }
        }
      },
      "GradeItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "date": {
            "type": "string",
            "example": "2022-12-01 17:53:52"
          },
          "grade": {
            "type": "integer",
            "example": 5
          },
          "description": {
            "type": "string",
            "example": "Webprogramozás I."
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "username"
          },
          "password": {
            "type": "string",
            "example": "password"
          }
        }
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "integer",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "Successfully logged in"
          },
          "token": {
            "type": "string",
            "example": "1111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111"
          }
        }
      },
      "RegisterResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "integer",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "Successfully registered!"
          }
        }
      },
      "Drawing_data": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 4
          },
          "username": {
            "type": "string",
            "example": "admin"
          },
          "title": {
            "type": "string",
            "example": "Teszt rajz"
          },
          "drawing_data": {
            "type": "string",
            "example": "1256893157125689315712568931571256893157125689315712568931571256893157125689315712568931571256893157"
          }
        }
      },
      "NewDrawing": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "c4fc72d6-68d9-4a7d-966e-fffa1d6cdffc"
          },
          "title": {
            "type": "string",
            "example": "Rajz címe"
          },
          "drawing_data": {
            "type": "string",
            "example": "1256893157125689315712568931571256893157125689315712568931571256893157125689315712568931571256893157"
          }
        }
      },
      "NewDrawingSuccess": {
        "type": "object",
        "properties": {
          "error": {
            "type": "integer",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "Drawing successfully uploaded!"
          }
        }
      },
      "DeleteDrawingResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "integer",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "Drawing successfully deleted!"
          }
        }
      },
      "DeleteDrawing": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "111111-111111-11111"
          },
          "id": {
            "type": "integer",
            "example": 5
          }
        }
      },
      "UpdateDrawingResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "integer",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "Drawing successfully updated!"
          }
        }
      },
      "UpdateDrawing": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "111111-111111-11111"
          },
          "title": {
            "type": "string",
            "example": "Teszt"
          },
          "drawing_data": {
            "type": "string",
            "example": "11111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111"
          }
        }
      },
      "SingleDrawing": {
        "type": "object",
        "properties": {
          "error": {
            "type": "integer",
            "format": "int64",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "Drawing data:"
          },
          "data": {
            "$ref": "#/components/schemas/Drawing_data"
          }
        }
      },
      "AllDrawings": {
        "type": "object",
        "properties": {
          "error": {
            "type": "integer",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "Listing drawings"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Drawing_data"
            }
          }
        }
      }
    }
  }
}